change unit_price type from float to decimal

+++ app/routers/offer.py
@@ -3,7 +3,7 @@ from fastapi_pagination import Page
 from fastapi_pagination.ext.sqlalchemy import paginate
 from sqlalchemy import func, or_
 from sqlalchemy.orm import Session
-from pydantic import Field
+from decimal import Decimal
 from database import get_db
 from models import Offer, Subcategory
 from schemas.offer import OfferResponse
@@ -41,9 +41,9 @@ If you choose category_id and subcategory_id, it will find offers matching to su
 def get_offers(
     category_id: int    | None = None,
     subcategory_id: int | None = None,
-    min_price: float    | None = Query(default=None, gt=0),
-    max_price: float    | None = Query(default=None, gt=0),
-    is_active: bool    | None = None,
+    min_price: Decimal  | None = Query(default=None, gt=0, decimal_places=2),
+    max_price: Decimal  | None = Query(default=None, gt=0, decimal_places=2),
+    is_active: bool     | None = None,
     db: Session = Depends(get_db)) -> Page[OfferResponse]:
 
     result = filter_offers_by_active_status(db, is_active)
